requirements_cuda128.txt
# ============================================================================
# PyTorch with CUDA Support
# ============================================================================
# NOTE: PyTorch should be installed FIRST via setup_env_cuda128.sh script
# Do NOT install torch directly from this requirements file
# The setup script ensures the correct CUDA 12.8 build is installed

# ============================================================================
# Core Scientific Computing
# ============================================================================
numpy>=1.24.0,<2.1.0
pandas>=2.0.0
scipy>=1.11.0

# ============================================================================
# Deep Learning Infrastructure
# ============================================================================
pytorch-lightning>=2.0.0
hydra-core>=1.3.0

# ============================================================================
# Megatron-Bridge and CUDA Extensions
# ============================================================================
nvidia-ml-py>=13.580.65
megatron-bridge==0.2.0rc6
# transformer_engine[pytorch] is installed explicitly in setup_env_cuda128.sh

# ============================================================================
# NLP and Tokenization
# ============================================================================
transformers>=4.35.0
sentencepiece>=0.1.99
tiktoken>=0.5.0
python-levenshtein>=0.21.0

# ============================================================================
# Data Processing and ML Datasets
# ============================================================================
datasets>=2.14.0

# ============================================================================
# Experiment Tracking and Logging
# ============================================================================
wandb>=0.15.0

# ============================================================================
# Evaluation and Metrics
# ============================================================================
mteb>=2.2.0
prettytable>=3.8.0
scikit-learn>=1.3.0

# ============================================================================
# Visualization and Analysis
# ============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================================
# Jupyter Notebooks
# ============================================================================
ipykernel>=7.0.0
jupyter>=1.0.0

# ============================================================================
# Data Handling
# ============================================================================
openpyxl>=3.1.0

# ============================================================================
# Development Tools
# ============================================================================
pre-commit>=3.4.0
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
